Move Mapper to separate 'properties' module
Leave in the main module empty migration aid class
